# services/backend/app/api/worlds.py
import base64
import binascii
import hashlib
import json
import logging
import math
import mimetypes
import shutil
from uuid import uuid4

from fastapi import APIRouter, HTTPException, Request, Response, WebSocket, WebSocketDisconnect
from fastapi.responses import FileResponse
from ..services.worlds import check, now, segment, validate_graph, world_graph, snap, node_ref, compute_route

router = APIRouter()
logger = logging.getLogger(__name__)

# Newest image queries kept per world; older JPEGs are deleted with their records.
LOCALIZATIONS_KEPT = 50
QUERY_IMAGE_MAX_BYTES = 2 * 1024 * 1024


async def body(request):
    raw = bytearray()
    async for chunk in request.stream():
        raw.extend(chunk)
        if len(raw) > 8*1024*1024:
            raise HTTPException(413, 'JSON body exceeds 8 MiB')
    try:
        data = json.loads(raw)
    except (ValueError, UnicodeDecodeError):
        raise HTTPException(400, 'Invalid JSON')
    if not isinstance(data, dict):
        raise HTTPException(400, 'Expected object')
    return data


@router.get('/worlds')
async def worlds(request: Request):
    store = request.app.state.worlds
    values = []
    root = store.path('worlds')
    for path in sorted(root.glob('*/world.json')):
        try:
            values.append(store.world(path.parent.name))
        except Exception:
            logger.warning('Skipped invalid world manifest: %s', path.parent.name)
    return {'worlds': values}


@router.post('/worlds', status_code=201)
async def create_world(request: Request):
    data = await body(request)
    if not {'id', 'name'} <= data.keys() or data.keys()-{'id', 'name', 'space', 'description', 'nianticSiteId'}:
        raise HTTPException(400, 'Expected id, name and optional world metadata')
    if not isinstance(data['id'], str):
        raise HTTPException(400, 'Invalid world ID')
    segment(data['id'])
    world = check({'schema': 'wander.world/v1', **data, 'nianticSiteId': data.get('nianticSiteId'),
        'version': 'v1', 'status': 'draft', 'assets': {'splat': f"worlds/{data['id']}/v1/scene.spz"},
        'updatedAt': now()}, 'world.schema.json')
    store = request.app.state.worlds
    async with store.lock('world:' + world['id']):
        if store.path('worlds', world['id'], 'world.json').exists():
            raise HTTPException(409, 'World already exists')
        await store.write(world, 'worlds', world['id'], 'world.json')
    return world


@router.get('/worlds/{world_id}')
async def get_world(world_id: str, request: Request):
    return request.app.state.worlds.world(world_id)


@router.patch('/worlds/{world_id}')
async def patch_world(world_id: str, request: Request):
    data = await body(request)
    if data.keys()-{'name', 'space', 'description', 'nianticSiteId', 'version', 'alignment', 'status', 'stats'}:
        raise HTTPException(400, 'Unsupported manifest field')
    store = request.app.state.worlds
    async with store.lock('world:' + world_id):
        original = store.world(world_id)
        world = check({**original, **data, 'updatedAt': now()}, 'world.schema.json')
        if 'version' in data:
            directory = store.path('worlds', world_id, world['version'])
            if not directory.is_dir():
                raise HTTPException(400, 'Asset version directory does not exist')
            world['assets'] = {key: f"worlds/{world_id}/{world['version']}/{value.split('/')[-1]}"
                               for key, value in original['assets'].items()
                               if (directory / value.split('/')[-1]).is_file()}
            if 'splat' not in world['assets']:
                splats = [p for p in directory.iterdir() if p.is_file() and p.suffix.lower() in ('.spz', '.ply', '.splat', '.ksplat', '.sog')]
                if len(splats) != 1:
                    raise HTTPException(400, 'Version needs one unambiguous splat asset')
                world['assets']['splat'] = f"worlds/{world_id}/{world['version']}/{splats[0].name}"
            for key, filename in [('mesh', 'mesh.glb'), ('vpsMap', 'vps-map.bin'), ('thumbnail', 'thumbnail.png')]:
                if (directory / filename).is_file():
                    world['assets'][key] = f"worlds/{world_id}/{world['version']}/{filename}"
        await store.write(world, 'worlds', world_id, 'world.json')
    return world


@router.delete('/worlds/{world_id}', status_code=204)
async def delete_world(world_id: str, request: Request):
    store = request.app.state.worlds
    async with store.lock('world:' + world_id):
        store.world(world_id)
        # path() verifies components, containment and absence of symlinks first.
        shutil.rmtree(store.path('worlds', world_id))
        await store.flush()
    return Response(status_code=204)


@router.put('/worlds/{world_id}/graph')
async def graph(world_id: str, request: Request):
    data = await body(request)
    validate_graph(data)
    store = request.app.state.worlds
    async with store.lock('world:' + world_id):
        world = store.world(world_id)
        world.update(navigationGraph=data, updatedAt=now())
        await store.write(world, 'worlds', world_id, 'world.json')
    return world


@router.get('/worlds/{world_id}/measurements')
async def measurements(world_id: str, request: Request):
    store = request.app.state.worlds
    store.world(world_id)
    if not store.path('worlds', world_id, 'measurements.json').exists():
        return {'schema': 'wander.measurements/v1', 'worldId': world_id, 'measurements': []}
    return check(store.read('worlds', world_id, 'measurements.json'), 'measurements.schema.json')


@router.put('/worlds/{world_id}/measurements')
async def put_measurements(world_id: str, request: Request):
    data = check(await body(request), 'measurements.schema.json')
    if data['worldId'] != world_id:
        raise HTTPException(400, 'Measurement worldId mismatch')
    store = request.app.state.worlds
    async with store.lock('world:' + world_id):
        store.world(world_id)
        data['updatedAt'] = now()
        await store.write(data, 'worlds', world_id, 'measurements.json')
    return data


@router.post('/worlds/{world_id}/route')
async def route(world_id: str, request: Request):
    return compute_route(request.app.state.worlds.world(world_id), await body(request))


@router.put('/worlds/{world_id}/annotations')
async def annotations(world_id: str, request: Request):
    """Backend extension; canonical manifest stays compliant with main's schema."""
    from ..services.annotations import AnnotationBatch, ReviewFile, publish_world
    data = await body(request)
    if set(data) != {'batch', 'review'}:
        raise HTTPException(400, 'Expected batch and review')
    batch = AnnotationBatch.model_validate(data['batch'])
    review = ReviewFile.model_validate(data['review'])
    store = request.app.state.worlds
    async with store.lock('world:' + world_id):
        manifest, records = publish_world(batch, review, store.world(world_id))
        path = store.path('worlds', world_id, 'annotations.json')
        previous = store.read('worlds', world_id, 'annotations.json') if path.exists() else []
        ids = {c.id for c in batch.candidates}
        records = [r for r in previous if r['id'] not in ids] + records
        await store.write(records, 'worlds', world_id, 'annotations.json')
        await store.write(manifest, 'worlds', world_id, 'world.json')
    return manifest


@router.post('/worlds/{world_id}/index')
async def index_world(world_id: str, request: Request):
    store = request.app.state.worlds
    async with store.lock('world:' + world_id):
        manifest = store.world(world_id)
        records = store.read('worlds', world_id, 'annotations.json') if store.path('worlds', world_id, 'annotations.json').exists() else []
        await request.app.state.elastic.setup()
        await request.app.state.ingestion.world(manifest, records)
    return {'indexed': len(world_graph(manifest)['nodes'])}


@router.post('/sessions', status_code=201)
async def create_session(request: Request):
    data = await body(request)
    if not {'worldId', 'deviceId'} <= data.keys() or data.keys()-{'worldId', 'deviceId', 'destination'}:
        raise HTTPException(400, 'Expected worldId, deviceId and optional destination')
    if not all(isinstance(v, str) and v for v in data.values()):
        raise HTTPException(400, 'Session fields must be nonempty strings')
    return await request.app.state.world_navigation.create(data['worldId'], data['deviceId'], data.get('destination'))


@router.get('/sessions/{session_id}')
async def session(session_id: str, request: Request):
    return request.app.state.worlds.session(session_id)


@router.delete('/sessions/{session_id}', status_code=204)
async def end_session(session_id: str, request: Request):
    store = request.app.state.worlds
    async with store.lock('session:' + session_id):
        data = store.session(session_id)
        data.update(state='ended', updatedAt=now())
        await store.save_session(data)
        await store.emit(data, 'ended')
    return Response(status_code=204)


@router.put('/sessions/{session_id}/destination')
async def destination(session_id: str, request: Request):
    data = await body(request)
    if set(data) != {'destination'} or not isinstance(data['destination'], str):
        raise HTTPException(400, 'Expected destination node ID')
    return await request.app.state.world_navigation.destination(session_id, data['destination'])


@router.post('/sessions/{session_id}/pose')
async def pose(session_id: str, request: Request):
    return await request.app.state.world_navigation.pose(session_id, await body(request))


@router.post('/worlds/{world_id}/localize')
async def localize(world_id: str, request: Request):
    data = check(await body(request), 'navigation.schema.json', '#/$defs/localizationUpdate')
    store, navigation = request.app.state.worlds, request.app.state.world_navigation
    world = store.world(world_id)
    if not world['nianticSiteId'] or data['nianticSiteId'] != world['nianticSiteId']:
        raise HTTPException(409, 'Niantic site mismatch')
    # alignment is splat->world, NOT VPS->world. Do not double-transform a VPS pose.
    if world.get('alignment', {}).get('frame') != 'niantic-vps':
        raise HTTPException(409, 'VPS localization requires alignment.frame=niantic-vps; no VPS-to-world transform is defined')
    nearest, (distance, position, _, _) = snap(world_graph(world), data['pose']['position'])
    if 'sessionId' in data:
        session = store.session(data['sessionId'])
        if session['worldId'] != world_id or session['deviceId'] != data['deviceId']:
            raise HTTPException(409, 'Session world/device mismatch')
    else:
        session = await navigation.create(world_id, data['deviceId'])
    await navigation.pose(session['sessionId'], {k: data[k] for k in ('pose', 'timestamp', 'trackingState') if k in data},
                          allow_no_destination=True)
    if data.get('trackingState', 'localized') == 'localized':
        async with store.lock('world:' + world_id):
            await store.write({'lastLocalizedAt': data['timestamp'], 'nianticSiteId': data['nianticSiteId']},
                              'worlds', world_id, 'vps-status.json')
        await store.emit(store.session(session['sessionId']), 'localized')
    return {'sessionId': session['sessionId'], 'worldId': world_id, 'pose': data['pose'],
            'nearestNode': {**node_ref(nearest), 'distanceMetres': math.dist(data['pose']['position'], nearest['position'])},
            'snappedPosition': position, 'offGraphMetres': distance}


def localizations_index(store, world_id):
    if store.path('worlds', world_id, 'localizations', 'index.json').exists():
        return check(store.read('worlds', world_id, 'localizations', 'index.json'),
                     'navigation.schema.json', '#/$defs/localizationQueries')
    return {'schema': 'wander.localizations/v1', 'worldId': world_id, 'queries': []}


@router.post('/worlds/{world_id}/localize/query', status_code=201)
async def localize_query(world_id: str, request: Request):
    """One VPS image query mirrored from the phone: the frame the SDK sent, its request record, the pose it produced."""
    data = check(await body(request), 'navigation.schema.json', '#/$defs/localizationQueryUpload')
    store = request.app.state.worlds
    world = store.world(world_id)
    if not world['nianticSiteId'] or data['nianticSiteId'] != world['nianticSiteId']:
        raise HTTPException(409, 'Niantic site mismatch')
    try:
        image = base64.b64decode(data['imageBase64'], validate=True)
    except (binascii.Error, ValueError):
        raise HTTPException(400, 'imageBase64 is not valid base64')
    if len(image) > QUERY_IMAGE_MAX_BYTES:
        raise HTTPException(413, 'Query image exceeds 2 MiB')
    if image[:3] != b'\xff\xd8\xff':
        raise HTTPException(400, 'Query image must be a JPEG')
    if 'sessionId' in data:
        session = store.session(data['sessionId'])
        if session['worldId'] != world_id or session['deviceId'] != data['deviceId']:
            raise HTTPException(409, 'Session world/device mismatch')

    query_id = 'q-' + uuid4().hex
    record = {'schema': 'wander.localization-query/v1', 'id': query_id, 'worldId': world_id,
              **{k: data[k] for k in ('sessionId', 'deviceId', 'role', 'nianticSiteId', 'capturedAt', 'request', 'result') if k in data},
              'receivedAt': now(),
              'image': {**data['image'], 'path': f'worlds/{world_id}/localizations/{query_id}.jpg'}}
    pose = data['result'].get('pose')
    if pose is not None and world.get('alignment', {}).get('frame') == 'niantic-vps':
        try:
            nearest, (distance, _, _, _) = snap(world_graph(world), pose['position'])
            record['nearestNode'] = {**node_ref(nearest), 'distanceMetres': math.dist(pose['position'], nearest['position'])}
            record['offGraphMetres'] = distance
        except HTTPException as error:
            if error.status_code != 409:  # a world without waypoints is fine here
                raise
    check(record, 'navigation.schema.json', '#/$defs/localizationQuery')

    async with store.lock('localizations:' + world_id):
        await store.write_bytes(image, 'worlds', world_id, 'localizations', query_id + '.jpg')
        index = localizations_index(store, world_id)
        kept, evicted = index['queries'][:LOCALIZATIONS_KEPT - 1], index['queries'][LOCALIZATIONS_KEPT - 1:]
        for old in evicted:
            store.path('worlds', world_id, 'localizations', old['id'] + '.jpg').unlink(missing_ok=True)
        index.update(queries=[record] + kept, updatedAt=record['receivedAt'])
        await store.write(index, 'worlds', world_id, 'localizations', 'index.json')
    if data['result']['trackingState'] == 'localized':
        async with store.lock('world:' + world_id):
            await store.write({'lastLocalizedAt': data['capturedAt'], 'nianticSiteId': data['nianticSiteId']},
                              'worlds', world_id, 'vps-status.json')
    return record


@router.get('/worlds/{world_id}/localizations')
async def localizations(world_id: str, request: Request, limit: int = 20):
    """Recent image queries, newest first. Images are served as `/worlds/{id}/localizations/{queryId}.jpg`."""
    store = request.app.state.worlds
    store.world(world_id)
    index = localizations_index(store, world_id)
    index['queries'] = index['queries'][:max(1, min(limit, LOCALIZATIONS_KEPT))]
    return index


@router.get('/worlds/{world_id}/vps')
async def vps(world_id: str, request: Request):
    store = request.app.state.worlds
    world = store.world(world_id)
    if not world['nianticSiteId']:
        raise HTTPException(409, 'World has no Niantic site')
    # False means not verified activated. A map export or old fix is not activation proof.
    result = {'nianticSiteId': world['nianticSiteId'], 'localizable': False, 'checkedAt': now(),
              'meshAvailable': False}
    asset = world['assets'].get('vpsMap')
    if asset:
        result['meshAvailable'] = store.path(*asset.split('/')).is_file()
    if store.path('worlds', world_id, 'vps-status.json').exists():
        status = store.read('worlds', world_id, 'vps-status.json')
        if status['nianticSiteId'] == world['nianticSiteId']:
            result['lastLocalizedAt'] = status['lastLocalizedAt']
    return result


@router.get('/worlds/{world_id}/{version}/{filename}')
async def asset(world_id: str, version: str, filename: str, request: Request):
    store = request.app.state.worlds
    store.world(world_id)
    path = store.path('worlds', world_id, version, filename)
    if not path.is_file():
        raise HTTPException(404, 'Asset not found')
    return FileResponse(path, media_type=mimetypes.guess_type(filename)[0] or 'application/octet-stream',
                        headers={'Cache-Control': 'public, max-age=31536000, immutable'})


@router.put('/worlds/{world_id}/{version}/{filename}', status_code=201)
async def upload(world_id: str, version: str, filename: str, request: Request):
    store = request.app.state.worlds
    path = store.path('worlds', world_id, version, filename)
    if path.suffix.lower() not in ('.spz', '.ply', '.splat', '.ksplat', '.sog', '.glb', '.png', '.bin'):
        raise HTTPException(400, 'Unsupported asset format')
    async with store.lock('world:' + world_id):
        store.world(world_id)
        if path.exists():
            raise HTTPException(409, 'Versioned asset already exists')
        path.parent.mkdir(parents=True, exist_ok=True)
        temporary = path.with_name('upload-' + uuid4().hex)
        digest, size = hashlib.sha256(), 0
        try:
            with temporary.open('xb') as file:
                async for chunk in request.stream():
                    size += len(chunk)
                    if size > request.app.state.settings.asset_max_bytes:
                        raise HTTPException(413, 'Asset exceeds configured limit')
                    digest.update(chunk)
                    file.write(chunk)
            temporary.replace(path)
            await store.flush()
        finally:
            temporary.unlink(missing_ok=True)
    return {'path': f'worlds/{world_id}/{version}/{filename}', 'bytes': size, 'sha256': digest.hexdigest()}


@router.websocket('/ws/sessions/{session_id}')
async def events(socket: WebSocket, session_id: str):
    store = socket.app.state.worlds
    try:
        session = store.session(session_id)
    except HTTPException:
        await socket.close(code=1008)
        return
    await socket.accept()
    clients = store.sockets.setdefault(session_id, set())
    clients.add(socket)
    try:
        kind = session['state'] if session['state'] in ('arrived', 'lost', 'ended') else 'progress'
        event = {'type': kind, 'sessionId': session_id, 'timestamp': now()}
        for source, target in [('lastPose', 'pose'), ('lastProgress', 'progress'), ('route', 'route')]:
            if source in session:
                event[target] = session[source]
        await socket.send_json(event)
        while True:
            raw = await socket.receive_text()
            try:
                valid = len(raw) <= 128 and json.loads(raw) == {'type': 'ping'}
            except ValueError:
                valid = False
            if not valid:
                await socket.close(code=1008, reason='Only ping messages are supported')
                return
    except WebSocketDisconnect:
        pass
    finally:
        clients.discard(socket)